# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Splite API",
    "version": "2.0.0-contract",
    "description": "Bill splitting for Venezuelan restaurants.\n\nSettlement is always VES in céntimos. USD and EUR are display references, never payment currencies.\n\n## Wire format conventions\n\nEvery endpoint follows these rules. A frontend that trusts them never needs to guess.\n\n| Kind | JSON type | Format | Example |\n|------|-----------|--------|---------|\n| **Money** (minor units) | `string` | Digit string — no leading zeros except `\"0\"` | `\"9007199254740993\"` |\n| **FX rates** | `string` | Decimal, padded to 8 fractional digits | `\"757.54060000\"` |\n| **IDs** | `string` | UUID v4 | `\"d290f1ee-6c54-4b01-90e6-d701748f0851\"` |\n| **Timestamps** | `string` | ISO 8601 `date-time` | `\"2025-03-05T16:30:00.000Z\"` |\n| **Value dates** | `string` | ISO 8601 `date` | `\"2025-03-06\"` |\n\nAmounts are strings so values beyond 2^53 survive JSON (JavaScript `Number` loses\nprecision past `Number.MAX_SAFE_INTEGER`). Rates are strings so every endpoint returns\nthe same representation — no `757.5406` from one route and `\"757.54060000\"` from another.\n\nEvery query is scoped to the caller's restaurant. A resource belonging to another tenant\nis reported as 404 rather than 403, so an endpoint never confirms that it exists.",
    "x-error-details": {
      "VALIDATION_FAILED": {
        "fields": "string[] -- one entry per failed field"
      },
      "OPEN_BILL_EXISTS": {
...
  "servers": [
...
  "security": [
    {
      "staffAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "staffAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived staff access token from /api/v1/auth/login."
      },
      "guestAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Guest session token. Send the token as a bearer credential *and* the session id in the `X-Guest-Session` header; both are required. Obtained from POST /api/v1/guest/sessions by presenting a signed table QR."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, malformed or expired credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected failure. The message is never echoed; correlate using requestId.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/guest/bill": {
      "get": {
        "tags": [
          "Guest"
        ],
        "summary": "The open bill for the guest's own table",
        "operationId": "getGuestBill",
        "description": "Authenticated with a guest session: `X-Guest-Session` plus the guest token as a bearer.\n\n**Takes no bill id.** The table comes from the session, which came from a signed QR, so\na guest cannot request a bill that is not theirs -- there is no identifier to tamper with.\n\nReturns 404 when the table has no open bill, which is the normal state between sittings.\n\nRate limited to 30 requests a minute per IP.",
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The bill.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestBill"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/v1/bills": {
      "get": {
        "tags": [
          "Bills"
        ],
        "summary": "List bills",
        "description": "Any authenticated staff role. Scoped to the caller's restaurant.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "CLOSED",
                "VOID"
              ]
            }
          },
          {
            "name": "tableId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bills.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Open a bill for a table",
        "x-required-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER",
          "WAITER"
        ],
        "description": "Roles: OWNER, MANAGER, CASHIER, WAITER. A table may have only one OPEN bill; a second attempt returns 409 `OPEN_BILL_EXISTS` with the existing bill id.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bill opened.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/bills/tables/{tableId}/open": {
      "get": {
        "tags": [
          "Bills"
        ],
        "summary": "Resolve a table to its current open bill",
        "description": "What a client scanning a permanent table QR calls. Any authenticated staff role.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TableId"
          }
        ],
        "responses": {
          "200": {
            "description": "The open bill.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/bills/{id}": {
      "get": {
        "tags": [
          "Bills"
        ],
        "summary": "Read a bill",
        "description": "Any authenticated staff role.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BillId"
          }
        ],
        "responses": {
          "200": {
            "description": "The bill, with its line items.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillWithItems"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/bills/{id}/items": {
      "get": {
        "tags": [
          "Bills"
        ],
        "summary": "List the lines on a bill",
        "description": "Any authenticated staff role.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BillId"
          }
        ],
        "responses": {
          "200": {
            "description": "Lines.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillItemList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Add a line to a bill",
        "x-required-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER",
          "WAITER"
        ],
        "description": "Roles: OWNER, MANAGER, CASHIER, WAITER.\n\nThe product name and price are **snapshotted** onto the line, so later menu changes\nnever alter a bill already served. Adding the same product twice creates two lines: a\nsecond round may have been ordered at a different price.\n\nThe bill total is recomputed from its lines and re-converted at the rate **frozen when\nthe bill opened**, never at the current rate.\n\nOnly an OPEN bill accepts lines. A bill opened with a fixed non-zero total is refused\nwith 409 `BILL_NOT_ITEMISED`; open it with a total of 0 to itemise it.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BillId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBillItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Line added.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillItemMutation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/bills/{id}/items/{itemId}": {
      "patch": {
        "tags": [
          "Bills"
        ],
        "summary": "Change a line quantity",
        "x-required-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER",
          "WAITER"
        ],
        "description": "Roles: OWNER, MANAGER, CASHIER, WAITER. The snapshotted unit price is never revisited; only the quantity changes.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BillId"
          },
          {
            "$ref": "#/components/parameters/BillItemId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBillItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillItemMutation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Bills"
        ],
        "summary": "Remove a line from a bill",
        "x-required-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER",
          "WAITER"
        ],
        "description": "Roles: OWNER, MANAGER, CASHIER, WAITER.\n\nRefused with 409 `TOTAL_BELOW_AMOUNT_PAID` when it would drop the bill total below what\nhas already been settled — reversing money that has moved is a refund, not an edit.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BillId"
          },
          {
            "$ref": "#/components/parameters/BillItemId"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillItemRemoval"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/v1/bills/{id}/void": {
      "post": {
        "tags": [
          "Bills"
        ],
        "summary": "Void an unpaid bill",
        "x-required-roles": [
          "OWNER",
          "MANAGER"
        ],
        "description": "Roles: OWNER, MANAGER. Refused once any payment has been applied: reversing money that has moved is a refund, not a status change.",
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/BillId"
          }
        ],
        "responses": {
          "200": {
            "description": "Voided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}